+++ src/com/yanncebron/m68kplugin/inspections/M68kDirectivesInspection.java
@@ -38,6 +38,11 @@ public class M68kDirectivesInspection extends LocalInspectionTool {
     return new M68kVisitor() {
 
       @Override
+      public void visitAutoDirective(@NotNull M68kAutoDirective element) {
+        holder.registerProblem(element, M68kBundle.message("inspection.directives.unsupported"));
+      }
+
+      @Override
       public void visitEndDirective(@NotNull M68kEndDirective element) {
         final M68kPsiElement nextSibling = PsiTreeUtil.getNextSiblingOfType(element, M68kPsiElement.class);
         if (nextSibling != null) {
@@ -87,10 +92,7 @@ public class M68kDirectivesInspection extends LocalInspectionTool {
     };
   }
 
-  private static void checkUnmatchedDirectiveWithNesting(M68kDirective m68kDirective, Class<? extends M68kDirective> directiveClass,
-                                                         ProblemsHolder holder, boolean forwards,
-                                                         Class<? extends M68kDirective> matchingDirective,
-                                                         String matchingDirectiveText) {
+  private static void checkUnmatchedDirectiveWithNesting(M68kDirective m68kDirective, Class<? extends M68kDirective> directiveClass, ProblemsHolder holder, boolean forwards, Class<? extends M68kDirective> matchingDirective, String matchingDirectiveText) {
     Ref<Boolean> unmatched = Ref.create(Boolean.TRUE);
     final Processor<M68kPsiElement> nestingProcessor = new Processor<>() {
 
@@ -125,28 +127,18 @@ public class M68kDirectivesInspection extends LocalInspectionTool {
 
   @SuppressWarnings("UnusedReturnValue")
   @SafeVarargs
-  private static void checkUnmatchedOpeningDirective(M68kDirective m68kDirective, ProblemsHolder holder,
-                                                     Class<? extends M68kDirective> matchingDirective,
-                                                     @NonNls String matchingDirectiveText,
-                                                     Class<? extends M68kDirective>... stopAtDirectives) {
+  private static void checkUnmatchedOpeningDirective(M68kDirective m68kDirective, ProblemsHolder holder, Class<? extends M68kDirective> matchingDirective, @NonNls String matchingDirectiveText, Class<? extends M68kDirective>... stopAtDirectives) {
     checkUnmatchedDirective(m68kDirective, holder, true, matchingDirective, matchingDirectiveText, stopAtDirectives);
   }
 
   @SuppressWarnings("UnusedReturnValue")
   @SafeVarargs
-  private static void checkUnmatchedClosingDirective(M68kDirective m68kDirective, ProblemsHolder holder,
-                                                     Class<? extends M68kDirective> matchingDirective,
-                                                     @NonNls String matchingDirectiveText,
-                                                     Class<? extends M68kDirective>... stopAtDirectives) {
+  private static void checkUnmatchedClosingDirective(M68kDirective m68kDirective, ProblemsHolder holder, Class<? extends M68kDirective> matchingDirective, @NonNls String matchingDirectiveText, Class<? extends M68kDirective>... stopAtDirectives) {
     checkUnmatchedDirective(m68kDirective, holder, false, matchingDirective, matchingDirectiveText, stopAtDirectives);
   }
 
   @SafeVarargs
-  private static void checkUnmatchedDirective(M68kDirective m68kDirective, ProblemsHolder holder,
-                                              boolean forwards,
-                                              Class<? extends M68kDirective> matchingDirective,
-                                              @NonNls String matchingDirectiveText,
-                                              Class<? extends M68kDirective>... stopAtDirectives) {
+  private static void checkUnmatchedDirective(M68kDirective m68kDirective, ProblemsHolder holder, boolean forwards, Class<? extends M68kDirective> matchingDirective, @NonNls String matchingDirectiveText, Class<? extends M68kDirective>... stopAtDirectives) {
     if (forwards) {
       if (M68kPsiTreeUtil.hasSiblingForwards(m68kDirective, matchingDirective, stopAtDirectives)) return;
     } else {
@@ -156,8 +148,7 @@ public class M68kDirectivesInspection extends LocalInspectionTool {
     addUnmatchedProblem(m68kDirective, holder, forwards, matchingDirectiveText);
   }
 
-  private static void addUnmatchedProblem(M68kDirective m68kDirective, ProblemsHolder holder,
-                                          boolean forwards, @NonNls String matchingDirectiveText) {
+  private static void addUnmatchedProblem(M68kDirective m68kDirective, ProblemsHolder holder, boolean forwards, @NonNls String matchingDirectiveText) {
     final String key = forwards ? "highlight.unmatched.directive.missing.closing" : "highlight.unmatched.directive.missing.opening";
     holder.registerProblem(m68kDirective, M68kBundle.message(key, matchingDirectiveText));
   }