+++ src/com/yanncebron/m68kplugin/lang/psi/M68kTokenGroups.java
@@ -172,6 +172,7 @@ public final class M68kTokenGroups {
     AC68080,
     ADDWATCH,
     ALIGN,
+    AUTO,
     BLK,
     BSS,
     BSS_C,